requirements.txt
Flask==1.1.2
numpy==1.19.2
pandas==1.1.3
scikit-learn==0.23.2
tensorflow==2.3.1
keras==2.4.3
nltk==3.5
matplotlib==3.3.2
seaborn==0.11.0
cryptography==3.1.1
bcrypt==3.2.0
pyjwt==1.7.1
flask-cors==3.0.9
flask-jwt-extended==3.24.1
flask-sqlalchemy==2.4.4
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.23.1
flask-migrate==2.5.3
gunicorn==20.0.4
pytest==6.1.1
coverage==5.3
flake8==3.8.4
black==20.8b1
isort==5.6.4
pre-commit==2.7.1
selenium==3.141.0
beautifulsoup4==4.9.3
requests==2.24.0
python-dotenv==0.15.0
pylint==2.6.0
mypy==0.790
